Add files via upload
The first game I made in C language! -> "TETRIS"
This game is known as: Tetris, bricks, falling blocks, etc.
In this game, pieces with different shapes and sizes appear on the screen and fall down, the goal is for the player to arrange the pieces to fill a line, then that line is removed and added to the player's score. . If the pieces rise and reach the top of the board, the player loses.